Make enemies turn when moving, implement collision
When enemies are attacking, instead of just using a stearing force, they
will now rotate to try and face the player and use a velocity forward
relatice to their rotation.

When an enemy collides with the player, it calls a function on the
player to do damage. The function is currently empty.