Implement enemy shooting, fix bug in enemy returning
Durring the first phase of attacking, enemies shoot every second. The
projectiles call the damage function on the player when they hit.

Sometimes when an enemy is returning to the formation, it would break
and rapidly rotate. This has been averted (not entierly fixed) by
clamping enemy rotation and allowing them to rotate faster.